+++ lib/src/service/compliance/job_type.dart
@@ -1,0 +1,10 @@
+// Copyright 2022 Kato Shinya. All rights reserved.
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided the conditions.
+
+/// This represents the compliance job type.
+enum JobType {
+  tweets,
+
+  users,
+}