+++ lib/src/twitter_api.dart
@@ -5,6 +5,7 @@
 // Project imports:
 import 'package:twitter_api_v2/src/client/client_context.dart';
 import 'package:twitter_api_v2/src/client/oauth_tokens.dart';
+import 'package:twitter_api_v2/src/service/compliance/compliance_service.dart';
 import 'package:twitter_api_v2/src/service/lists/lists_service.dart';
 import 'package:twitter_api_v2/src/service/spaces/spaces_service.dart';
 import 'package:twitter_api_v2/src/service/tweets/tweets_service.dart';
@@ -30,6 +31,9 @@ abstract class TwitterApi {
 
   /// Returns the lists service.
   ListsService get listsService;
+
+  /// Returns the compliance service.
+  ComplianceService get complianceService;
 }
 
 class _TwitterApi implements TwitterApi {
@@ -57,4 +61,7 @@ class _TwitterApi implements TwitterApi {
 
   @override
   ListsService get listsService => _twitterService.listsService;
+
+  @override
+  ComplianceService get complianceService => _twitterService.complianceService;
 }